refactor(BEB-34): ajusta importacao das dependencias dos modulos
