+++ src/modules/atendees/atendee.module.ts
@@ -1,7 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AuthService } from 'src/shared/auth/services/auth.service';
-import { BcryptService } from 'src/shared/bcrypt/bcrypt.service';
+import { AuthModule } from 'src/shared/auth/auth.module';
 
 import { AuthAtendeeController, CreateAtendeeController } from './controllers';
 import { Atendees } from './entities';
@@ -9,14 +8,8 @@ import { AtendeeRepository } from './repositories';
 import { AuthAtendeeService, CreateAtendeeService } from './services';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Atendees])],
-  providers: [
-    AtendeeRepository,
-    AuthService,
-    AuthAtendeeService,
-    CreateAtendeeService,
-    BcryptService,
-  ],
+  imports: [TypeOrmModule.forFeature([Atendees]), AuthModule],
+  providers: [AtendeeRepository, AuthAtendeeService, CreateAtendeeService],
   controllers: [AuthAtendeeController, CreateAtendeeController],
 })
 export class AtendeeModule {}
