+++ src/modules/atendees/services/auth-atendee.service.ts
@@ -5,7 +5,7 @@ import {
 } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
 
-import { BcryptService } from '../../../shared/bcrypt/bcrypt.service';
+import { BcryptService } from '../../../shared/auth/services/bcrypt.service';
 import { AuthAtendeeInput } from '../input';
 import { AtendeeRepository } from '../repositories';
