+++ src/modules/users/user.module.ts
@@ -1,8 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { BcryptService } from 'src/shared/bcrypt/bcrypt.service';
 
-import { AuthService } from '../../shared/auth/services/auth.service';
+import { AuthModule } from '../../shared/auth/auth.module';
 import { AuthUserController, CreateUserController } from './controllers';
 import { Users } from './entities';
 import { UserRepository } from './repositories';
@@ -13,15 +12,14 @@ import {
 } from './services';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Users])],
+  imports: [TypeOrmModule.forFeature([Users]), AuthModule],
   providers: [
     CreateUserService,
     ReadUserByIdService,
     UserRepository,
-    BcryptService,
     AuthUserService,
-    AuthService,
   ],
   controllers: [CreateUserController, AuthUserController],
+  exports: [ReadUserByIdService],
 })
 export class UserModule {}
