+++ src/shared/auth/auth.module.ts
@@ -1,19 +1,19 @@
 import { Module } from '@nestjs/common';
 import { JwtModule } from '@nestjs/jwt';
 
-import { UserModule } from '../../modules/users/user.module';
 import { jwtConstants } from '../constants/jwt.constants';
 import { AuthService } from './services/auth.service';
+import { BcryptService } from './services/bcrypt.service';
 
 @Module({
   imports: [
-    UserModule,
     JwtModule.register({
       global: true,
       secret: jwtConstants.secret,
       signOptions: { expiresIn: '172800s' },
     }),
   ],
-  providers: [AuthService],
+  providers: [AuthService, BcryptService],
+  exports: [AuthService, BcryptService],
 })
 export class AuthModule {}
